refactor(tasks): move tasks to separate files
