+++ tasks/copy.js
@@ -1,0 +1,12 @@
+'use strict';
+
+module.exports = function(gulp, config) {
+  return function(onlyChanged) {
+    var gulpif = require('gulp-if');
+    var changed = require('gulp-changed');
+
+    return gulp.src(config.server.filePattern)
+      .pipe(gulpif(onlyChanged, changed(config.build.distPath)))
+      .pipe(gulp.dest(config.build.distPath));
+  };
+};
