+++ tasks/start.js
@@ -1,0 +1,24 @@
+'use strict';
+
+module.exports = function(gulp, config) {
+  return function() {
+    var nodemon = require('gulp-nodemon');
+    var notifier = require('node-notifier');
+    var path = require('path');
+
+    nodemon({
+      script: config.server.runnable,
+      ext: 'js jade',
+      watch: [config.build.distPath],
+      delay: 1,
+      env: config.server.environmentVariables,
+      nodeArgs: ['--harmony']
+    }).on('restart', function() {
+      notifier.notify({
+        title: 'Boar tasks',
+        message: 'Server restarted',
+        icon: path.join(__dirname, '..', 'boar.png')
+      });
+    });
+  };
+};
